Fixes memory leak when normalizing the host os.
For reasons not yet understood, ActiveRecord is not releasing the strings
generated by the Enumerable "some_host.services." I don't know if this
is unique to services. This was confirmed by inspecting the ObjectSpace
of the Ruby process during an import of an Nmap XML file wherein every
port on a few hosts are listed as open.

This problem is still ongoing, because there's no way I'm going to
abandon each()'ing all AR collections, so this fix should be considered
a stopgap.

For background, see https://gist.github.com/2721110