Clarify reasoning for restricting version < 1.0.0
[#44034071]

One of the previous reasons for not going to 1.0.0 was that the
migrations were in metasploit-framework, but with metasploit_data_models
now having a migration not in metasploit-framework
(20130228214900_change_required_columns_to_null_false_in_web_vulns.rb),
metasploit_data_models now must be the canonical source and
metasploit-framework's Msf::DbManager#migrate will have to look to
MetasploitDataModels.root.join('db', 'migrate') for its default
migrations path instead of data/sql/migrate.  This means that the
reasoning in lib/metasploit_data_models/version.rb needs to remove that
it is because the migrations aren't really being used in
metasploit_data_models by other projects and instead its because the
migrations are not tested in specs.