Mdm::Host#module_details
[#47979793]

Pro needs to access the modules that were run against a given host.
This was previously done through a convoluted join across the refs.name
and module_refs.name.  Instead of having to remember this odd path, add
deep has_many throughs to Mdm::Host and has_many over the name join for
Mdm::Module::Ref and Mmd::Ref, so that Mdm::Host#module_details can be
used to list the modules run on a host directly.

Additional changes come as side-effects and giving mice cookies: adding
yard docs for the new associations led to links to undocumented object
warnings, so that spawned more docs.  Not wanting the documentation to
be incomplete for the affected classes, I wrote all the docs for the
classes that were missing.  To test the long chain of has_many throughs,
factories for all the intermediate models needed to be made, so there
are now factories and specs to check those factories are valid.