Required columns non-null in module_actions
[#50752269]

Forgot that MetasploitDataModels::ChangeRequiredColumnsToNullFalse
existed, so use that for new ChangeColumnNullInModuleArchs and change
ChangeColumnNullInModuleActions to use it also.  Move deletion of NULL
rows (which had a bug) from ChangeColumnNullInModuleActions to
ChangeRequiredColumnsToNullFalse as its a general requirement to delete
rows in with NULLs before the columns can be made :null => false.