Mdm::Module::Ancestor and Mdm::Module::Class specs passing
[#50752269]

THIS COMMIT HAS FAILING SPECS.

This commit checks in the work for the past week or so now that the
database model has stabilized, but only Mdm::Module::Ancestor and
Mdm::Module::Class specs are passing.  Mdm::Module::Instance and all its
assocations, which used to be associations of Mdm::Module::Detail need
to be speced or have their specs fixed.