Move failure on undocumented to rake task
[#50752269]

Instead of travis-ci having to parse the rake yard output for the
percentage documented, I made the rake yard:stats task fail if the
undocumented count is > 0.  This means the error condition is checked
locally and handles if there were so many objects that the calculation
showed 100.00% documented to 2 decimal places, but there were still
undocumented objects.