architecture and platform association for Mdm::Session
MSP-9285

Mdm::Session#platform is removed as a String attribute in favor of
Mdm::Session#architecture and Mdm::Session#platform associations to
Mdm::Architecture and Mdm::Platform, respectively.  The associations are
validated for presence so that new sessions require them, but the
foreign key columns are null: true to support legacy sessions.