require 'bundler/gem_tasks' before 'bundler/setup'
MSP-11929

Attempt to work-around issue with building when there is no bundler
cache on travis-ci that is not seen on metasploit-cache.  First
difference was missing `require 'bundler/gem_tasks'` in
metasploit_data_models.  Also removed guard as
`Bundler::GemTask.install_tasks` would never have worked if bundler was
not installed, so the guard wasn't guarding properly and therefore it
was never being used.