Collapse nested namespaces
MSP-12489

Having lexically nested modules leads to the outer Modules being defined
in mutliple places so that the first file that defines the Module, such
as a child Class/Module's file, can stop the actual Module file being
loaded.  This results in a partially define Module/Class for the outer
namespaces and hard to diagnose loading bugs.